source/statistics/core/requirements.txt
Flask>=1.0.2
pytest>=4.6.2
flask-inputs>=0.3.0
jsonschema>=3.0.1
twisted
requests
chardet2
urllib3
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
mysqlclient
python-decouple
validators
marshmallow